Make components (BestItems, AllItems) and Add them to App.js
